requirements.txt
matplotlib
networkx
tqdm
beautifulsoup4
tree_sitter
tree_sitter_language_pack
sentence-transformers
faiss-cpu
numpy
pyyaml
pathspec
symspellpy
transformers
google-genai #if using a gemini API key (free API accses as of 7/20/2025)